requirements.txt
numpy==1.22.0
opencv-contrib-python==4.5.5.62
Pillow==9.0.0
tinydb==4.5.2
